Remove special CheckEquals function for intptr_t on X64 platform. Review URL: http://codereview.chromium.org/190004
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@2842 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/objects.h
@@ -2634,8 +2634,8 @@ class Code: public HeapObject {
   // the layout of the code object into account.
   int ExecutableSize() {
     // Check that the assumptions about the layout of the code object holds.
-    ASSERT_EQ(instruction_start() - address(),
-              static_cast<intptr_t>(Code::kHeaderSize));
+    ASSERT_EQ(static_cast<int>(instruction_start() - address()),
+              Code::kHeaderSize);
     return instruction_size() + Code::kHeaderSize;
   }
